Fix Gui extractRenderState mixin target and guard gradle signing config
Move the WrapOperation injection for overlay screen rendering from
GameRendererMixin#extractGui to FabricGuiMixin#extractRenderState, since
MC 26.2 moved the Screen.extractRenderStateWithTooltipAndSubtitles call
site from GameRenderer into Gui. Also only register the Maven signing
task when a signing key is actually configured, so publishToMavenLocal
does not fail looking for a non-existent .asc artifact.
