+++ buildSrc/src/main/groovy/multiloader-common.gradle
@@ -156,10 +156,12 @@ publishing {
 }
 
 signing {
+    // Only register the sign task when a key is actually configured, otherwise
+    // publishToMavenLocal fails looking for a non-existent .asc artifact.
     if (System.getenv("SIGNING_KEY")) {
         def signingKey = System.getenv("SIGNING_KEY")
         def signingPassword = System.getenv("SIGNING_PASSPHRASE")
         useInMemoryPgpKeys(signingKey, signingPassword)
+        sign publishing.publications.mavenJava
     }
-    sign publishing.publications.mavenJava
 }
